tripoille test now all ok

+++ ft_print_bigx.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_print_bigX.c                                    :+:      :+:    :+:   */
+/*   ft_print_bigx.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: sdukic <sdukic@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/21 18:44:32 by sdukic            #+#    #+#             */
-/*   Updated: 2022/10/21 21:20:38 by sdukic           ###   ########.fr       */
+/*   Updated: 2022/10/23 18:06:06 by sdukic           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 #include "ft_printf.h"
 #include "libft/libft.h"
 
-int	ft_print_bigx(unsigned long int arg)
+int	ft_print_bigx(unsigned int arg)
 {
 	ft_putnbr_base(arg, "0123456789ABCDEF");
 	return (ft_numlen_base(arg, 16));
