+++ ft_print_x.c
@@ -6,7 +6,7 @@
 /*   By: sdukic <sdukic@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/21 18:44:32 by sdukic            #+#    #+#             */
-/*   Updated: 2022/10/21 21:20:37 by sdukic           ###   ########.fr       */
+/*   Updated: 2022/10/23 18:05:00 by sdukic           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 #include "ft_printf.h"
 #include "libft/libft.h"
 
-int	ft_print_x(unsigned long int arg)
+int	ft_print_x(unsigned int arg)
 {
 	ft_putnbr_base(arg, "0123456789abcdef");
 	return (ft_numlen_base(arg, 16));
